Correction of the lexical analyzer.
The variables part of the symbol table is now filled. It was not because of the modification of the declaration of a variable. Previously, we had 'a s9(5)' for example, but now it's more like 'a pic s9(5)'.
Because of the architecture of the code, the variable[0] variable was reset to null at each 'pic' recognized, thus avoiding the addition of a variable entry in the symbols table.